fix: fix style of home page section loading
