+++ site/src/components/home/sections/latestBlocks.js
@@ -1,11 +1,8 @@
 import { Flex, FlexBetween } from "../../styled/flex";
-import { Panel } from "../../styled/panel";
 import styled from "styled-components";
 import {
   Inter_12_500,
-  Inter_14_500,
   Inter_14_600,
-  Inter_18_700,
   SF_Mono_14_500,
 } from "../../../styles/text";
 import { withLoading } from "../../../HOC/withLoading";
@@ -14,6 +11,7 @@ import { timeDuration } from "../../../utils/viewFuncs/time";
 import { ReactComponent as Block } from "./block.svg";
 import { addressEllipsis } from "../../../utils/viewFuncs";
 import Link from "../../styled/link";
+import Loading from "../../loadings/loading";
 
 const BlockIcon = styled(Block)`
   path {
@@ -31,16 +29,11 @@ const BlockIcon = styled(Block)`
   }
 `;
 
-const Title = styled.h2`
-  ${Inter_18_700};
-  color: ${(props) => props.theme.fontPrimary};
-`;
-
 const Rows = styled.ul`
   margin: 0;
   padding-left: 0;
   padding-top: 8px;
-  max-width: 500px;
+  max-width: 644px;
   display: flex;
   flex-wrap: wrap;
 `;
@@ -91,60 +84,57 @@ const Label = styled.span`
   color: ${(props) => props.theme.fontTertiary};
 `;
 
-const Anchor = styled(Link)`
-  display: block;
-  padding-right: 24px;
-  ${Inter_14_500};
-  line-height: 52px;
-  text-align: right;
-  color: ${(p) => p.theme.theme500};
+const LoadingContainer = styled.div`
+  width: 100vw;
+  margin-left: calc(-50vw + 50%);
 `;
 
 const mapLoadingState = (props) => {
   const { blocks } = props;
-  return { loadingStates: [blocks?.length === 0] };
+  return {
+    loadingStates: [blocks?.length === 0],
+    loadingComponent: (
+      <LoadingContainer>
+        <Loading />
+      </LoadingContainer>
+    ),
+  };
 };
 
 function LatestBlocks({ blocks }) {
   return (
-    <div>
-      <Title>Latest Blocks</Title>
-      <Panel>
-        <Rows>
-          {blocks.slice(0, 5).map((block, i) => (
-            <Row key={i}>
-              <FlexBetween>
-                <Flex style={{ gap: 16 }}>
-                  <BlockIcon />
-                  <div>
-                    <Link to={`/block/${block.height}`}>
-                      {" "}
-                      <ThemeText>{block.height}</ThemeText>
-                    </Link>
-                    <Time> {timeDuration(block.time)} </Time>
-                  </div>
-                </Flex>
+    <Rows>
+      {blocks.slice(0, 5).map((block, i) => (
+        <Row key={i}>
+          <FlexBetween>
+            <Flex style={{ gap: 16 }}>
+              <BlockIcon />
+              <div>
+                <Link to={`/block/${block.height}`}>
+                  {" "}
+                  <ThemeText>{block.height}</ThemeText>
+                </Link>
+                <Time> {timeDuration(block.time)} </Time>
+              </div>
+            </Flex>
 
-                <div>
-                  <Address>{addressEllipsis(block.validator)}</Address>
-                  <Flex style={{ fontSize: 12, gap: 16 }}>
-                    <Flex style={{ gap: 8 }}>
-                      <Label>Extrinsics</Label>
-                      <Bold>{block.extrinsicsCount}</Bold>
-                    </Flex>
-                    <Flex style={{ gap: 8 }}>
-                      <Label>Events</Label>
-                      <Bold>{block.eventsCount}</Bold>{" "}
-                    </Flex>
-                  </Flex>
-                </div>
-              </FlexBetween>
-            </Row>
-          ))}
-        </Rows>
-        <Anchor to={"/blocks"}>View All</Anchor>
-      </Panel>
-    </div>
+            <div>
+              <Address>{addressEllipsis(block.validator)}</Address>
+              <Flex style={{ fontSize: 12, gap: 16 }}>
+                <Flex style={{ gap: 8 }}>
+                  <Label>Extrinsics</Label>
+                  <Bold>{block.extrinsicsCount}</Bold>
+                </Flex>
+                <Flex style={{ gap: 8 }}>
+                  <Label>Events</Label>
+                  <Bold>{block.eventsCount}</Bold>{" "}
+                </Flex>
+              </Flex>
+            </div>
+          </FlexBetween>
+        </Row>
+      ))}
+    </Rows>
   );
 }
